Create InferenceSession and moved InferenceInput to inference_session
